refactor(hooks): split into module directory
