+++ src/hooks/mod.rs
@@ -1,0 +1,9 @@
+mod context;
+mod point;
+mod resolve;
+mod runner;
+
+pub use context::HookContext;
+pub use point::HookPoint;
+pub use resolve::{resolve_hook, resolve_on_failure};
+pub use runner::run_hook;
